+++ app/src/main/java/woowacourse/movie/view/home/complete/BookingCompleteActivity.kt
@@ -16,9 +16,9 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.databinding.DataBindingUtil
 import androidx.room.Room
 import woowacourse.movie.R
-import woowacourse.movie.data.TicketDatabase
-import woowacourse.movie.data.TicketEntity.Companion.TICKET_TABLE_NAME
-import woowacourse.movie.data.toEntity
+import woowacourse.movie.data.ticket.TicketDatabase
+import woowacourse.movie.data.ticket.TicketEntity.Companion.TICKET_TABLE_NAME
+import woowacourse.movie.data.ticket.toEntity
 import woowacourse.movie.databinding.ActivityBookingCompleteBinding
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.domain.model.ticket.Ticket
