+++ app/src/main/java/woowacourse/movie/view/setting/SettingFragment.kt
@@ -15,6 +15,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import woowacourse.movie.R
+import woowacourse.movie.data.setting.SettingStorageManagerImpl
 import woowacourse.movie.databinding.FragmentSettingBinding
 
 class SettingFragment :
